1. 글 작성시 작성자에 닉네임 들어가도록 수정 2. 작성일자 들어가도록 수정 3. 조회수 0으로 초기화 4. 제목 클릭 시 마다 조회수 올라가도록 수정

+++ src/main/java/sjspring/shop/pregAndBirthDeveloper/dto/BoardListViewResponse.java
@@ -11,7 +11,7 @@ public class BoardListViewResponse {
     private final String content;
     private final String author;
     private final LocalDateTime createdAt;
-    private final Long views;
+    private final int views;
 
     public BoardListViewResponse(Board board){
         this.boardNo = board.getBoardNo();
